sample_chat/requirements.txt
fastapi
pydantic
uvicorn
psycopg2
sqlalchemy
email-validator
passlib
python-jose[cryptography]